+++ src/frontend/components/DashboardVideoLiveControlPane/index.tsx
@@ -1,19 +1,20 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 
+import { DashboardVideoLiveWidgetsContainer } from 'components/DashboardVideoLiveControlPane/widgets/DashboardVideoLiveWidgetsContainer';
+import { DashboardVideoLiveWidgetGeneralTitle } from 'components/DashboardVideoLiveControlPane/widgets/DashboardVideoLiveWidgetGeneralTitle';
 import { Video } from 'types/tracks';
-import { DashboardVideoLiveWidgetsContainer } from './widgets/DashboardVideoLiveWidgetsContainer';
 
 interface DashboardVideoLiveControlPaneProps {
   video: Video;
 }
-/** This component will hold all the widgets in the futur, and will be displayed by DashboardVideoLive component */
+
 export const DashboardVideoLiveControlPane = ({
   video,
 }: DashboardVideoLiveControlPaneProps) => {
-  video = video;
   return (
     <DashboardVideoLiveWidgetsContainer>
-      {[]}
+      <DashboardVideoLiveWidgetGeneralTitle video={video} />
+      <Fragment />
     </DashboardVideoLiveWidgetsContainer>
   );
 };